moved logging functions from utils to cli_logging
