+++ bia-ingest/bia_ingest/ingest/image_annotation_dataset.py
@@ -1,11 +1,12 @@
 import logging
 from typing import List, Any, Dict, Optional
+
+from ..cli_logging import log_model_creation_count
 from .utils import (
     find_sections_recursive,
     dicts_to_api_models,
     dict_to_uuid,
     filter_model_dictionary,
-    log_model_creation_count,
 )
 import bia_ingest.ingest.study as study_conversion
 from .biostudies import (
