+++ bia-ingest/bia_ingest/ingest/specimen_growth_protocol.py
@@ -1,11 +1,12 @@
 import logging
 from typing import List, Any, Dict, Optional
+
+from ..cli_logging import log_model_creation_count
 from .utils import (
     dicts_to_api_models,
     find_sections_recursive,
     dict_to_uuid,
     filter_model_dictionary,
-    log_model_creation_count,
 )
 from .biostudies import (
     Submission,
