+++ bia-ingest/bia_ingest/representation_creation/experimentally_captured_image.py
@@ -3,12 +3,12 @@ from uuid import UUID
 from typing import List, Dict, Any
 from pydantic import ValidationError
 
+from ..cli_logging import log_failed_model_creation, log_model_creation_count
+
 from .utils import merge_dicts
 from ..ingest.utils import (
     dict_to_uuid,
     filter_model_dictionary,
-    log_model_creation_count,
-    log_failed_model_creation,
 )
 from ..ingest.biostudies import (
     Submission,
